requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.8.1
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
colorama==0.4.6
cssselect2==0.8.0
distro==1.9.0
Django==5.2.3
django-environ==0.12.0
fonttools==4.60.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
openai==1.12.0
pillow==11.3.0
psycopg2-binary==2.9.9
pycparser==2.23
pydantic==2.12.0
pydantic_core==2.41.1
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==1.0.0
setuptools==70.2.0
sniffio==1.3.1
sqlparse==0.5.0
tinycss2==1.4.0
tinyhtml5==2.0.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2024.1
weasyprint==65.1
webencodings==0.5.1
wheel==0.37.1
zopfli==0.2.3.post1
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
django-cors-headers==4.9.0
